pyproject.toml
[tool.poetry]
name = "octavio_web"
version = "0.1.0"
description = ""
authors = ["occy88 <octavio.delser@gmail.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.11"
pydantic-settings = "^2.2.1"
pydantic = "^2.6.2"
pytz = "^2024.1"


[tool.poetry.group.test.dependencies]
pytest-asyncio = "^0.23.5"
pytest-xdist = "^3.5.0"
pytest-cov = "^4.1.0"
pytest = {extras = ["psycopg2"], version = "^8.0.1"}
pytest-mock = "^3.12.0"


[tool.poetry.group.web.dependencies]
sqlalchemy = "^2.0.27"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
celery = {extras = ["redis"], version = "^5.3.6"}
httpx = "^0.27.0"
sqlmodel = "^0.0.16"
sqlalchemy-utils = "^0.41.1"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
fastapi-crudrouter = "^0.8.6"
factory-boy = "^3.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
